requirements.txt
canonicalwebteam.flask_base==1.1.0
canonicalwebteam.blog==6.4.0
canonicalwebteam.discourse-docs==1.0.1
canonicalwebteam.http==1.0.3
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.search==1.3.0
canonicalwebteam.image-template==1.3.1
feedparser==6.0.10
requests-cache==0.9.7
lxml==4.9.2
cchardet==2.1.7; platform_machine != 'aarch64'
pyyaml==6.0
